requirements.txt
streamlit
pandas
openpyxl
curl_cffi
Pillow
python-pptx
beautifulsoup4